Change imageio as_gray to mode=L due to imageio v2->v3 (#739)
* Update ads_utils.py

* Point to IVADOMED torch update branch

* Convert some remaining imreads to our native  imread

* Fix pip install from github branch nomenclature

* Fix branch name

* Try editable pip install and commit instead of branch

* Remove temp files

* Remove editable flag

* Missing git+

* Fix bug due to implicit bool casting to system uint instead of uint8

* Remove 6+ year old  unused (and buggy) feature

* Update get_masks.py

* Typo fix

* Point ivadomed to latest master commit

---------

+++ AxonDeepSeg/visualization/generate_axons_from_myelin.py
@@ -38,7 +38,7 @@ def generate_axons_from_myelin(path_prediction,path_myelin_corrected):
     new_axon_mask = np.logical_xor(myelin_corrected, fused)
 
     # merge corrected myelin mask and generated axon mask
-    both = new_axon_mask*255 + myelin_corrected*127
+    both = new_axon_mask*np.uint8(255) + myelin_corrected*np.uint8(127)
 
     # save the corrected axon+myelin image
     path = path_prediction.parent / 'axon_myelin_mask_corrected.png'
